Release V1.0

+++ src/DxtImage.cpp
@@ -83,7 +83,7 @@ namespace scially {
         int imgSize = img->getImageSizeInBytes();
         int x_pos = 0;
         int y_pos = 0;
-        for (size_t i = 0; i < imgSize; i += 8)
+        for (int i = 0; i < imgSize; i += 8)
         {
             // 64 bit matrix
             unsigned short color0, color1;
